Android: fixed QAndroidVideoRendererControl.
When requesting a SurfaceTexture from this class to be used with the
media player or the camera, the scenegraph GL context might not be
available yet, resulting in the control not being able to share the
texture with it.
We now wait for the scenegraph GL context to be available before
creating our GL context and texture.

Change-Id: Iea6bf8af11d81b3ae0662d4d370cd6cc458e849f
Reviewed-by: Christian Stromme <christian.stromme@digia.com>